# swagger.yaml
# OpenAPI 3.0 specification for Task API
openapi: 3.0.0

info:
  title: Task API Test
  version: 1.0.0
  description: >
    A RESTful Task API built with Node.js, Express.js, PostgreSQL, and Supabase Auth.
    The API supports CRUD operations for tasks and user authentication.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:

  # =========================
  # API INFORMATION
  # =========================

  /:
    get:
      summary: API Information
      description: Returns basic information about the Task API.
      responses:
        "200":
          description: API information returned successfully


  # =========================
  # HEALTH CHECK
  # =========================

  /health:
    get:
      summary: Health Check
      description: Checks whether the API is running.
      responses:
        "200":
          description: API is healthy


  # =========================
  # AUTHENTICATION
  # =========================

  /auth/signup:
    post:
      summary: Register a new user
      description: Creates a new user account using Supabase Authentication.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: test@example.com

                password:
                  type: string
                  format: password
                  example: password123

      responses:
        "201":
          description: User successfully registered

        "400":
          description: Email or password is missing or registration failed


  /auth/login:
    post:
      summary: Log in a user
      description: Authenticates a user using Supabase Authentication and returns access and refresh tokens.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: test@example.com

                password:
                  type: string
                  format: password
                  example: password123

      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                    example: eyJhbGciOiJIUzI1NiIs...

                  refresh_token:
                    type: string
                    description: Token used to refresh the session
                    example: abc123...

        "400":
          description: Email or password is missing

        "401":
          description: Invalid login credentials



  /auth/logout:
    post:
      summary: Log out the current user
      description: Logs out the authenticated user. Requires a valid access token.

      security:
        - bearerAuth: []

      responses:
        "204":
          description: Logout successful

        "401":
          description: Access token required or token is invalid/expired

        "500":
          description: Logout failed
    # =========================
  # PUBLIC & PROTECTED ROUTES
  # =========================

  /public/info:
    get:
      summary: Public information
      description: Public endpoint that does not require authentication.
      responses:
        "200":
          description: Public information returned successfully


  /protected/profile:
    get:
      summary: Protected profile
      description: Protected endpoint that requires a Bearer access token.
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Access token in the format Bearer <token>
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIs...
      
      security:
        - bearerAuth: []

      responses:
        "200":
          description: Protected profile accessed successfully

        "401":
          description: Access token required

      #protected route for dashboard
  /protected/dashboard:
    get:
      summary: Access protected dashboard
      description: Returns dashboard information for an authenticated user. Requires a valid access token.

      security:
        - bearerAuth: []

      responses:
        "200":
          description: Protected dashboard accessed successfully

        "401":
          description: Access token required or token is invalid/expired

        "500":
          description: Authentication error
  # =========================
  # TASK COLLECTION
  # =========================

  /tasks:
    get:
      summary: Get all tasks
      description: Returns all tasks stored in PostgreSQL.
      responses:
        "200":
          description: List of tasks returned successfully

    post:
      summary: Create a new task
      description: Creates a new task in PostgreSQL.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: Build Candidate API

                done:
                  type: integer
                  description: Task completion status. Use 0 for incomplete and 1 for completed.
                  example: 0

      responses:
        "201":
          description: Task created successfully

        "400":
          description: Title is required

        "500":
          description: Database error


  # =========================
  # INDIVIDUAL TASK
  # =========================

  /tasks/{id}:

    get:
      summary: Get a task by ID
      description: Returns a single task from PostgreSQL.

      parameters:
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
          example: 1

      responses:
        "200":
          description: Task found

        "404":
          description: Task not found

        "500":
          description: Database error


    put:
      summary: Update a task
      description: Updates an existing task in PostgreSQL.

      parameters:
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
          example: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Build Candidate API

                done:
                  type: integer
                  description: Task completion status. Use 0 for incomplete and 1 for completed.
                  example: 1

      responses:
        "200":
          description: Task updated successfully

        "404":
          description: Task not found

        "500":
          description: Database error


    delete:
      summary: Delete a task
      description: Deletes an existing task from PostgreSQL.

      parameters:
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
          example: 1

      responses:
        "204":
          description: Task deleted successfully

        "404":
          description: Task not found

        "500":
          description: Database error
    
    components:
      securitySchemes:
        bearerAuth:
          type: http
          scheme: bearer
          bearerFormat: JWT